requirements.txt
Pillow
piexif
opencv-python
numpy
